@@ -4,17 +4,26 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
+    <parent>
+        <groupId>org.jboss</groupId>
+        <artifactId>jboss-parent</artifactId>
+        <version>26</version>
+        <relativePath/>
+    </parent>
+
     <groupId>io.weli</groupId>
     <artifactId>Resteasy2168SpringBoot</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>1.8</java.version>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <resteasy.version>4.4.1.Final</resteasy.version>
-        <springboot.version>2.2.2.RELEASE</springboot.version>
+        <java.version>11</java.version>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <resteasy.version>6.2.1.Final</resteasy.version>
+        <springboot.version>3.0.0</springboot.version>
+        <version.resteasy.springboot>6.0.0.Final</version.resteasy.springboot>
+        <version.jakarta.ws.rs-api>3.1.0</version.jakarta.ws.rs-api>
     </properties>
 
     <dependencyManagement>
@@ -45,14 +54,14 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>javax.ws.rs</groupId>
-            <artifactId>javax.ws.rs-api</artifactId>
-            <version>2.1</version>
+            <groupId>jakarta.ws.rs</groupId>
+            <artifactId>jakarta.ws.rs-api</artifactId>
+            <version>${version.jakarta.ws.rs-api}</version>
         </dependency>
         <dependency>
             <groupId>org.jboss.resteasy</groupId>
-            <artifactId>resteasy-spring-boot-starter</artifactId>
-            <version>4.4.0.Final</version>
+            <artifactId>resteasy-servlet-spring-boot-starter</artifactId>
+            <version>${version.resteasy.springboot}</version>
             <scope>runtime</scope>
         </dependency>
     </dependencies>
@@ -63,7 +72,7 @@
                 <plugin>
                     <groupId>org.apache.maven.plugins</groupId>
                     <artifactId>maven-jar-plugin</artifactId>
-                    <version>3.1.0</version>
+<!--                    <version>3.1.0</version>-->
                 </plugin>
             </plugins>
         </pluginManagement>
